3rd motor added for drivetrain
Most of the code had two motors for each side missing the third motor they each had. i have put in the third motor into all the code except for the end where you look for the encoders. This is because i dont know if it's just finding the average or some other function you are using to find encoders.